Backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
python-dotenv>=1.0
SQLAlchemy>=2.0
psycopg2-binary>=2.9
pydantic>=2.6
email-validator>=2.1
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
python-jose[cryptography]>=3.3.0
mcp>=1.16.0
httpx>=0.27.0
aiofiles>=24.1.0
scholarly>=1.7.0
arxiv>=2.2.0
